+++ Modules/Bank/Http/Controllers/MaintenanceController.php
@@ -77,8 +77,6 @@ class MaintenanceController extends Controller
     {
         return Person::whereNull('police_no')
             ->whereNull('case_no_hash')
-            ->whereNull('registration_no')
-            ->whereNull('section_card_no')
             ->doesntHave('helper')
             ->count();
     }
@@ -122,8 +120,6 @@ class MaintenanceController extends Controller
         if (isset($request->cleanup_no_number)) {
             $cnt +=  Person::whereNull('police_no')
                 ->whereNull('case_no_hash')
-                ->whereNull('registration_no')
-                ->whereNull('section_card_no')
                 ->doesntHave('helper')
                 ->delete();
         }