+++ Modules/Bank/Http/Controllers/WithdrawalController.php
@@ -72,8 +72,6 @@ class WithdrawalController extends Controller
                     $aq->where(function($wq) use ($term) {
                        $wq->where('search', 'LIKE', '%' . $term . '%'); 
                        $wq->orWhere('police_no', $term);
-                       $wq->orWhere('registration_no', $term);
-                       $wq->orWhere('section_card_no', $term);
                        $wq->orWhere('case_no_hash', DB::raw("SHA2('". $term ."', 256)"));
                     });
                 }