+++ Modules/People/Database/Migrations/2019_09_22_121758_remove_section_registration_no_from_persons.php
@@ -1,0 +1,35 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class RemoveSectionRegistrationNoFromPersons extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('persons', function (Blueprint $table) {
+            $table->dropColumn('section_card_no');
+            $table->dropColumn('registration_no');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('persons', function (Blueprint $table) {
+            $table->string('registration_no')->nullable();
+            $table->string('section_card_no')->nullable();
+        });
+    }
+
+}