+++ Modules/People/Resources/views/index.blade.php
@@ -43,8 +43,6 @@
                         <th class="text-nowrap">@lang('people::people.date_of_birth') <a href="javascript:;" class="sort" data-field="date_of_birth">@icon(sort)</a></th>
                         <th>@lang('people::people.nationality') <a href="javascript:;" class="sort" data-field="nationality">@icon(sort)</a></th>
                         <th class="text-nowrap">@lang('people::people.police_no')</th>
-                        <th class="text-nowrap">@lang('people::people.reg_no')</th>
-                        <th class="text-nowrap">@lang('people::people.sec_card_no')</th>
                         <th>@lang('people::people.languages') <a href="javascript:;" class="sort" data-field="languages">@icon(sort)</a></th>
                         <th>@lang('people::people.remarks') <a href="javascript:;" class="sort" data-field="remarks">@icon(sort)</a></th>
                         <th rowspan="2" class="align-top">@lang('app.registered') <a href="javascript:;" class="sort" data-field="created_at">@icon(sort)</a></th>
@@ -55,8 +53,6 @@
                         <th>{{ Form::text('date_of_birth', !empty($filter['date_of_birth']) ? $filter['date_of_birth'] : null, [ 'class' => 'form-control form-control-sm', 'placeholder' => 'Filter...', 'autocomplete' => 'off' ]) }}</th>
                         <th>{{ Form::text('nationality', !empty($filter['nationality']) ? $filter['nationality'] : null, [ 'class' => 'form-control form-control-sm', 'placeholder' => 'Filter...', 'autocomplete' => 'off' ]) }}</th>
                         <th>{{ Form::text('police_no', !empty($filter['police_no']) ? $filter['police_no'] : null, [ 'class' => 'form-control form-control-sm', 'placeholder' => 'Filter...', 'autocomplete' => 'off' ]) }}</th>
-                        <th>{{ Form::text('registration_no', !empty($filter['registration_no']) ? $filter['registration_no'] : null, [ 'class' => 'form-control form-control-sm', 'placeholder' => 'Filter...', 'autocomplete' => 'off' ]) }}</th>
-                        <th>{{ Form::text('section_card_no', !empty($filter['section_card_no']) ? $filter['section_card_no'] : null, [ 'class' => 'form-control form-control-sm', 'placeholder' => 'Filter...', 'autocomplete' => 'off' ]) }}</th>
                         <th>{{ Form::text('languages', !empty($filter['languages']) ? $filter['languages'] : null, [ 'class' => 'form-control form-control-sm', 'placeholder' => 'Filter...', 'autocomplete' => 'off' ]) }}</th>
                         <th>{{ Form::text('remarks', !empty($filter['remarks']) ? $filter['remarks'] : null, [ 'class' => 'form-control form-control-sm', 'placeholder' => 'Filter...', 'autocomplete' => 'off' ]) }}</th>
                     </tr>