+++ Modules/People/Resources/views/show.blade.php
@@ -113,34 +113,6 @@
                         </div>
                     </li>
                 @endisset
-                @isset($person->registration_no)
-                    <li class="list-group-item">
-                        <div class="row">
-                            <div class="col-sm-5">
-                                <strong>
-                                    @lang('people::people.registration_number')
-                                </strong>
-                            </div>
-                            <div class="col-sm">
-                                {{ $person->registration_no }}
-                            </div>
-                        </div>
-                    </li>
-                @endisset
-                @isset($person->section_card_no)
-                    <li class="list-group-item">
-                        <div class="row">
-                            <div class="col-sm-5">
-                                <strong>
-                                    @lang('people::people.section_card_number')
-                                </strong>
-                            </div>
-                            <div class="col-sm">
-                                {{ $person->section_card_no }}
-                            </div>
-                        </div>
-                    </li>
-                @endisset
                 @isset($person->nationality)
                     <li class="list-group-item">
                         <div class="row">