+++ resources/js/people.js
@@ -50,8 +50,6 @@ $(function(){
         $('#filter input[name="date_of_birth"]').val('');
         $('#filter input[name="nationality"]').val('');
         $('#filter input[name="police_no"]').val('');
-        $('#filter input[name="registration_no"]').val('');
-        $('#filter input[name="section_card_no"]').val('');
         $('#filter input[name="languages"]').val('');
         $('#filter input[name="remarks"]').val('');
         filterTable(1);
@@ -82,8 +80,6 @@ function filterTable(page) {
         "date_of_birth": $('#filter input[name="date_of_birth"]').val(),
         "nationality": $('#filter input[name="nationality"]').val(),
         "police_no": $('#filter input[name="police_no"]').val(),
-        "registration_no": $('#filter input[name="registration_no"]').val(),
-        "section_card_no": $('#filter input[name="section_card_no"]').val(),
         "languages": $('#filter input[name="languages"]').val(),
         "remarks": $('#filter input[name="remarks"]').val(),
         "page": page,
@@ -179,8 +175,6 @@ function writeRow(person) {
         .append($('<td>').text(person.date_of_birth))
         .append($('<td>').text(person.nationality))
         .append($('<td>').text(person.police_no))
-        .append($('<td>').text(person.registration_no))
-        .append($('<td>').text(person.section_card_no))
         .append($('<td>').text(person.languages ? Array.isArray(person.languages) ? person.languages.join(', ') : person.languages : ''))
         .append($('<td>').text(person.remarks))
         // TODO .append($('<td>').html(person.helper ? '<i class="fa fa-check"></i>' : '-'))