+++ Changelog.md
@@ -7,6 +7,7 @@ Bank:
 * Added feature to block coupon types for newly registered users
 * Show daily limited coupons statistics in main bank screen
 * Only offer to create new person in bank if search results negative
+* Removed section card and registration number fields from person
 
 ## 2.7.6